fix tests, remove default schemas from processors and put them in the apis
