+++ asset/src/id_reader/slicer.ts
@@ -7,12 +7,14 @@ import ElasticsearchAPI from '../elasticsearch_reader_api/elasticsearch-api';
 import { ESIDReaderConfig } from './interfaces';
 import {
     ElasticReaderFactoryAPI,
-    IDSlicerArgs
+    IDSlicerArgs,
+    ElasticsearchReaderAPIConfig
 } from '../elasticsearch_reader_api/interfaces';
 
 export default class ESIDSlicer extends ParallelSlicer<ESIDReaderConfig> {
     api!: ElasticsearchAPI;
     version!: number;
+    config!: ElasticsearchReaderAPIConfig;
 
     async initialize(recoveryData: SlicerRecoveryData[]): Promise<void> {
         const apiName = this.opConfig.api_name;
@@ -23,9 +25,10 @@ export default class ESIDSlicer extends ParallelSlicer<ESIDReaderConfig> {
 
         const apiConfig = apiManager.getConfig(apiName);
         if (!apiConfig) throw new Error(`Could not find api config for api_name ${apiName}`);
+        this.config = apiConfig;
 
-        if (this.version !== 5 && apiConfig.field == null) {
-            throw new Error('Paramter field must be set if querying against elasticsearch version >= 6.x');
+        if (this.version <= 5 && apiConfig.id_field_name == null) {
+            throw new Error('Paramter id_field_name must be set if querying against elasticsearch version >= 6.x');
         }
         // NOTE ORDER MATTERS
         // a parallel slicer initialize calls newSlicer multiple times
@@ -40,7 +43,14 @@ export default class ESIDSlicer extends ParallelSlicer<ESIDReaderConfig> {
 
     async newSlicer(id: number): Promise<SlicerFn> {
         const { lifecycle, slicers } = this.executionConfig;
-        const { key_type, key_range, starting_key_depth } = this.opConfig;
+
+        const {
+            key_type,
+            key_range,
+            starting_key_depth,
+            id_field_name
+        } = this.config;
+
         const { recoveryData } = this;
 
         const args: IDSlicerArgs = {
@@ -48,9 +58,10 @@ export default class ESIDSlicer extends ParallelSlicer<ESIDReaderConfig> {
             numOfSlicers: slicers,
             slicerID: id,
             recoveryData,
-            key_type,
-            key_range: key_range || undefined,
-            starting_key_depth,
+            keyType: key_type,
+            keyRange: key_range || undefined,
+            startingKeyDepth: starting_key_depth,
+            IDFieldName: id_field_name || null
         };
 
         return this.api.makeIDSlicer(args as IDSlicerArgs);
