+++ asset/src/spaces_reader_api/schema.ts
@@ -4,9 +4,12 @@ import {
     getTypeOf,
     cloneDeep,
     isNumber,
+    ValidatedJobConfig
 } from '@terascope/job-components';
-import { schema } from '../elasticsearch_reader/schema';
+import elasticAPI from '@terascope/elasticsearch-api';
+import { schema } from '../elasticsearch_reader_api/schema';
 import { ApiConfig } from '../elasticsearch_reader/interfaces';
+import { DEFAULT_API_NAME } from './interfaces';
 
 const clone = cloneDeep(schema);
 
@@ -37,10 +40,19 @@ const spacesSchema = Object.assign({}, clone, apiSchema) as AnyObject;
 // this should not continue onward
 delete spacesSchema.api_name;
 
-spacesSchema.index.format = 'required_String';
 spacesSchema.date_field_name.format = 'required_String';
 
 export default class Schema extends ConvictSchema<ApiConfig> {
+    validateJob(job: ValidatedJobConfig): void {
+        const { logger } = this.context;
+
+        const apiConfigs = job.apis.filter((config) => config._name.startsWith(DEFAULT_API_NAME));
+
+        apiConfigs.forEach((apiConfig: AnyObject) => {
+            elasticAPI({}, logger).validateGeoParameters(apiConfig);
+        });
+    }
+
     build(): AnyObject {
         return spacesSchema;
     }
